refact: active api 호출 이벤트 useOper 훅으로 통합
#70
