--- src/components/rightSide/user/dropFunction.ts
@@ -1,8 +1,0 @@
-export function onProfile(
-  user: string,
-  setProfileUser: React.Dispatch<React.SetStateAction<string>>,
-  onClose: () => void,
-) {
-  setProfileUser(user);
-  onClose();
-}
